Move gamer colors to separate constants file.
