+++ app/src/main/java/noteworthyengine/units/City.java
@@ -1,5 +1,6 @@
 package noteworthyengine.units;
 
+import art.Constants;
 import noteworthyengine.BattleNode;
 import noteworthyengine.BattleSystem;
 import noteworthyengine.GridNode;
@@ -44,7 +45,7 @@ public class City extends Unit {
         battleNode.gamer.v = gamer;
         battleNode.hp.v = 110;
         battleNode.attackDamage.v = 1;
-        renderNode.set(0, 0, 0, 1.5f, 1.5f, 90f, Gamer.colorForTeam(gamer.team), "Animations/Buildings/City", 0, 0);
+        renderNode.set(0, 0, 0, 1.5f, 1.5f, 90f, Constants.colorForTeam(gamer.team), "Animations/Buildings/City", 0, 0);
     }
 
     @Override
