+++ app/src/main/java/noteworthyengine/units/Platoon.java
@@ -3,6 +3,7 @@ package noteworthyengine.units;
 import android.graphics.Color;
 
 import art.Animations;
+import art.Constants;
 import noteworthyengine.BattleNode;
 import noteworthyengine.BattleSystem;
 import noteworthyengine.FieldNode;
@@ -91,7 +92,7 @@ public class Platoon extends Unit {
 //                    (float)renderNode.coords.pos.x + 0.5f * (float)(battleNode.hp.v / 50), (float)renderNode.coords.pos.y + 0.7f, 2,
 //                    (battleNode.hp.v > 30) ? Color.GREEN : (battleNode.hp.v > 15) ? Color.YELLOW : Color.RED);
 
-            renderNode.color.v = Gamer.TeamColors.get(battleNode.gamer.v.team);
+            renderNode.color.v = Constants.colorForTeam(battleNode.gamer.v.team);
             //renderNode.color.v = Color.argb(10, 255, 255, 255);
 
             if (selectionNode.isSelected.v == 1) {
@@ -120,7 +121,7 @@ public class Platoon extends Unit {
                         Math.max(1, (float) (0.3 * Math.sin(time / 15)) + 0.9f),
                         Math.max(1, (float) (0.3 * Math.sin(time / 15)) + 0.9f),
                         time * 2,
-                        battleNode.gamer.v.color(),
+                        Constants.colorForTeam(battleNode.gamer.v.team),
                         RenderNode.RENDER_LAYER_FOREGROUND
                 );
             }
