+++ app/src/main/java/noteworthyengine/units/UnitPool.java
@@ -1,5 +1,6 @@
 package noteworthyengine.units;
 
+import art.Constants;
 import noteworthyframework.Gamer;
 import noteworthyframework.Unit;
 import utils.MemoryPool;
@@ -10,7 +11,7 @@ import utils.MemoryPool;
 public class UnitPool {
 
     public static final Gamer NO_GAMER = new Gamer("No gamer!") {{
-      this.team = Gamer.TeamColors.size() - 1;
+      this.team = Constants.TeamColors.size() - 1;
     }};
 
     public static final int NUMBER_UNITS = 256;
